requirements.txt
Django==5.1.3
djangorestframework==3.15.2
pytz==2024.2
pytest-django==4.9.0
model-bakery==1.20.0
